+++ waterfall-cad/CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to the
 [Haskell Package Versioning Policy](https://pvp.haskell.org/).
 
+
 ## Unreleased
+
+## 0.1.1.0 - 2023-12-12 
 
 ### Added